Refactoring of Console option management
The objective is to separate the part in charge of manage option for the application that the part in charge of manages options of the console.
That separation would make possible to change of HMI easely is the future

The command line system for reading option has been changed. Previously I made the choice to used Microsoft Configuration class to parth command lines options. But with this system it is not possible to retrieve option that has not argument.

A logger has also been added